Refactor extraction logic into ExtractionJob
