+++ ingestify/application/loader.py
@@ -1,20 +1,14 @@
-import itertools
-import json
 import logging
 import platform
 from multiprocessing import set_start_method
 from typing import List, Optional
 
-from ingestify.domain.models import Dataset, Identifier, Selector, Task, TaskSet
-from ingestify.utils import TaskExecutor, chunker
+from ingestify.domain.models import Selector
+from ingestify.utils import TaskExecutor
 
 from .dataset_store import DatasetStore
-from .. import DatasetResource, retrieve_http
-from ..domain import DraftFile
-from ingestify.domain.models.extraction.extraction_job_summary import ExtractionJobSummary
-from ingestify.domain.models.execution.extraction_plan import ExtractionPlan
-from ..domain.models.resources.dataset_resource import FileResource
-from ..domain.models.task.task_summary import TaskSummary
+from ingestify.domain.models.extraction.extraction_plan import ExtractionPlan
+from ..domain.models.extraction.extraction_job import ExtractionJob
 from ..exceptions import ConfigurationError
 
 if platform.system() == "Darwin":
@@ -26,135 +20,6 @@ else:
 logger = logging.getLogger(__name__)
 
 
-DEFAULT_CHUNK_SIZE = 1000
-
-
-def to_batches(input_):
-    if isinstance(input_, list):
-        batches = [input_]
-    else:
-        # Assume it's an iterator. Peek what's inside, and put it back
-        try:
-            peek = next(input_)
-        except StopIteration:
-            # Nothing to batch
-            return []
-
-        input_ = itertools.chain([peek], input_)
-
-        if not isinstance(peek, list):
-            batches = chunker(input_, DEFAULT_CHUNK_SIZE)
-        else:
-            batches = input_
-    return batches
-
-
-def load_file(
-    file_resource: FileResource, dataset: Optional[Dataset] = None
-) -> Optional[DraftFile]:
-    current_file = None
-    if dataset:
-        current_file = dataset.current_revision.modified_files_map.get(
-            file_resource.file_id
-        )
-
-    if file_resource.json_content is not None:
-        # Empty dictionary is allowed
-        file = DraftFile.from_input(
-            file_=json.dumps(file_resource.json_content, indent=4),
-            data_serialization_format="json",
-            data_feed_key=file_resource.data_feed_key,
-            data_spec_version=file_resource.data_spec_version,
-            modified_at=file_resource.last_modified,
-        )
-        if current_file and current_file.tag == file.tag:
-            # Nothing changed
-            return None
-        return file
-    elif file_resource.url:
-        http_options = {}
-        if file_resource.http_options:
-            for k, v in file_resource.http_options.items():
-                http_options[f"http_{k}"] = v
-
-        return retrieve_http(
-            url=file_resource.url,
-            current_file=current_file,
-            file_data_feed_key=file_resource.data_feed_key,
-            file_data_spec_version=file_resource.data_spec_version,
-            file_data_serialization_format=file_resource.data_serialization_format
-            or "txt",
-            **http_options,
-            **file_resource.loader_kwargs,
-        )
-    else:
-        return file_resource.file_loader(
-            file_resource,
-            current_file,
-            # TODO: check how to fix this with typehints
-            **file_resource.loader_kwargs,
-        )
-
-
-class UpdateDatasetTask(Task):
-    def __init__(
-        self,
-        dataset: Dataset,
-        dataset_resource: DatasetResource,
-        store: DatasetStore,
-    ):
-        self.dataset = dataset
-        self.dataset_resource = dataset_resource
-        self.store = store
-
-    def run(self):
-        with TaskSummary.update() as task_summary:
-            revision = self.store.update_dataset(
-                dataset=self.dataset,
-                dataset_resource=self.dataset_resource,
-                files={
-                    file_id: load_file(file_resource, dataset=self.dataset)
-                    for file_id, file_resource in self.dataset_resource.files.items()
-                },
-            )
-            task_summary.set_stats_from_revision(revision)
-            return task_summary
-
-    def __repr__(self):
-        return f"UpdateDatasetTask({self.dataset_resource.provider} -> {self.dataset_resource.dataset_resource_id})"
-
-
-class CreateDatasetTask(Task):
-    def __init__(
-        self,
-        dataset_resource: DatasetResource,
-        store: DatasetStore,
-    ):
-        self.dataset_resource = dataset_resource
-        self.store = store
-
-    def run(self):
-        with TaskSummary.create() as task_summary:
-            revision = self.store.create_dataset(
-                dataset_type=self.dataset_resource.dataset_type,
-                provider=self.dataset_resource.provider,
-                dataset_identifier=Identifier(**self.dataset_resource.dataset_resource_id),
-                name=self.dataset_resource.name,
-                state=self.dataset_resource.state,
-                metadata=self.dataset_resource.metadata,
-                files={
-                    file_id: load_file(file_resource)
-                    for file_id, file_resource in self.dataset_resource.files.items()
-                },
-            )
-
-            task_summary.set_stats_from_revision(revision)
-            return task_summary
-
-    def __repr__(self):
-        return f"CreateDatasetTask({self.dataset_resource.provider} -> {self.dataset_resource.dataset_resource_id})"
-
-
 class Loader:
     def __init__(self, store: DatasetStore):
         self.store = store
@@ -164,8 +29,6 @@ class Loader:
         self.extraction_plans.append(extraction_plan)
 
     def collect_and_run(self, dry_run: bool = False, provider: Optional[str] = None):
-        total_dataset_count = 0
-
         # First collect all selectors, before discovering datasets
         selectors = {}
         for extraction_plan in self.extraction_plans:
@@ -182,7 +45,9 @@ class Loader:
                 if not selector.is_dynamic
             ]
             dynamic_selectors = [
-                selector for selector in extraction_plan.selectors if selector.is_dynamic
+                selector
+                for selector in extraction_plan.selectors
+                if selector.is_dynamic
             ]
 
             no_selectors = len(static_selectors) == 1 and not bool(static_selectors[0])
@@ -239,17 +104,17 @@ class Loader:
             # Merge selectors when source, dataset_type and actual selector is the same. This makes
             # sure there will be only 1 dataset for this combination
             for selector in static_selectors:
-                key = (extraction_plan.source.name, extraction_plan.dataset_type, selector.key)
+                key = (
+                    extraction_plan.source.name,
+                    extraction_plan.dataset_type,
+                    selector.key,
+                )
                 if existing_selector := selectors.get(key):
                     existing_selector[1].data_spec_versions.merge(
                         selector.data_spec_versions
                     )
                 else:
                     selectors[key] = (extraction_plan, selector)
-
-        def run_task(task):
-            logger.info(f"Running task {task}")
-            return task.run()
 
         """
             Data is denormalized:
@@ -277,100 +142,14 @@ class Loader:
                 f"Discovering datasets from {extraction_plan.source.__class__.__name__} using selector {selector}"
             )
 
-            with ExtractionJobSummary(
-                extraction_plan=extraction_plan,
-                selector=selector
-            ) as extraction_job_summary:
+            extraction_job = ExtractionJob(
+                extraction_plan=extraction_plan, selector=selector
+            )
 
-                with extraction_job_summary.record_timing("get_dataset_collection"):
-                    dataset_collection_metadata = self.store.get_dataset_collection(
-                        dataset_type=extraction_plan.dataset_type,
-                        data_spec_versions=selector.data_spec_versions,
-                        selector=selector,
-                        metadata_only=True,
-                    ).metadata
-
-                # There are two different, but similar flows here:
-                # 1. The discover_datasets returns a list, and the entire list can be processed at once
-                # 2. The discover_datasets returns an iterator of batches, in this case we need to process each batch
-                with extraction_job_summary.record_timing("find_datasets"):
-                    # Timing might be incorrect as it is an iterator
-                    datasets = extraction_plan.source.find_datasets(
-                        dataset_type=extraction_plan.dataset_type,
-                        data_spec_versions=selector.data_spec_versions,
-                        dataset_collection_metadata=dataset_collection_metadata,
-                        **selector.custom_attributes,
-                    )
-
-                batches = to_batches(datasets)
-
-                with extraction_job_summary.record_timing("tasks"):
-                    for batch in batches:
-                        dataset_identifiers = [
-                            Identifier.create_from_selector(
-                                selector, **dataset_resource.dataset_resource_id
-                            )
-                            # We have to pass the data_spec_versions here as a Source can add some
-                            # extra data to the identifier which is retrieved in a certain data format
-                            for dataset_resource in batch
-                        ]
-
-                        # Load all available datasets based on the discovered dataset identifiers
-                        dataset_collection = self.store.get_dataset_collection(
-                            dataset_type=extraction_plan.dataset_type,
-                            # Assume all DatasetResources share the same provider
-                            provider=batch[0].provider,
-                            selector=dataset_identifiers,
-                        )
-
-                        skip_count = 0
-                        total_dataset_count += len(dataset_identifiers)
-
-                        task_set = TaskSet()
-                        for dataset_resource in batch:
-                            dataset_identifier = Identifier.create_from_selector(
-                                selector, **dataset_resource.dataset_resource_id
-                            )
-
-                            if dataset := dataset_collection.get(dataset_identifier):
-                                if extraction_plan.fetch_policy.should_refetch(
-                                    dataset, dataset_resource
-                                ):
-                                    task_set.add(
-                                        UpdateDatasetTask(
-                                            dataset=dataset,  # Current dataset from the database
-                                            dataset_resource=dataset_resource,  # Most recent dataset_resource
-                                            store=self.store,
-                                        )
-                                    )
-                                else:
-                                    skip_count += 1
-                            else:
-                                if extraction_plan.fetch_policy.should_fetch(dataset_resource):
-                                    task_set.add(
-                                        CreateDatasetTask(
-                                            dataset_resource=dataset_resource,
-                                            store=self.store,
-                                        )
-                                    )
-                                else:
-                                    skip_count += 1
-
-                        if task_set:
-                            logger.info(
-                                f"Discovered {len(dataset_identifiers)} datasets from {extraction_plan.source.__class__.__name__} "
-                                f"using selector {selector} => {len(task_set)} tasks. {skip_count} skipped."
-                            )
-                            logger.info(f"Running {len(task_set)} tasks")
-                            with TaskExecutor(dry_run=dry_run) as task_executor:
-                                extraction_job_summary.add_task_summaries(
-                                    task_executor.run(run_task, task_set)
-                                )
-                        else:
-                            logger.info(
-                                f"Discovered {len(dataset_identifiers)} datasets from {extraction_plan.source.__class__.__name__} "
-                                f"using selector {selector} => nothing to do"
-                            )
+            with TaskExecutor(dry_run=dry_run) as task_executor:
+                extraction_job_summary = extraction_job.execute(
+                    self.store, task_executor=task_executor
+                )
 
                 # TODO: handle task_summaries
                 #       Summarize to a ExtractionJobSummary, and save to a database. This Summary can later be used in a
