+++ ingestify/domain/models/extraction/extraction_job_summary.py
@@ -4,25 +4,16 @@ from datetime import datetime, timedelta
 from typing import Optional
 
 from ingestify.domain import Selector
-from ingestify.domain.models.execution.extraction_plan import ExtractionPlan
-from ingestify.domain.models.task.task_summary import TaskSummary
+from ingestify.domain.models.extraction.extraction_plan import ExtractionPlan
+from ingestify.domain.models.task.task_summary import TaskSummary, TaskStatus
+from ingestify.domain.models.timing import Timing
 from ingestify.utils import utcnow
 
 
 def format_duration(duration: timedelta):
     from humanize import naturaldelta
+
     return naturaldelta(duration, minimum_unit="MILLISECONDS")
-
-
-@dataclass
-class Timing:
-    name: str
-    start: datetime
-    end: datetime
-
-    @property
-    def duration(self):
-        return self.end - self.start
 
 
 @dataclass
@@ -39,15 +30,9 @@ class ExtractionJobSummary:
     def record_timing(self, name: str):
         start = utcnow()
         yield
-        self.timings.append(
-            Timing(
-                name=name,
-                start=start,
-                end=utcnow()
-            )
-        )
+        self.timings.append(Timing(name=name, start=start, end=utcnow()))
 
-    def add_task_summaries(self, task_summaries):
+    def add_task_summaries(self, task_summaries: list[TaskSummary]):
         self.task_summaries.extend(task_summaries)
 
     def set_finished(self):
@@ -58,14 +43,27 @@ class ExtractionJobSummary:
         return self.end - self.start
 
     def output_report(self):
-        print(f"\nExtractionJobSummary finished in {format_duration(self.end - self.start)}")
+        print(
+            f"\nExtractionJobSummary finished in {format_duration(self.end - self.start)}"
+        )
         print("--------------------")
         print(f"  - ExtractionPlan: {self.extraction_plan}")
         print(f"  - Selector: {self.selector}")
         print(f"  - Timings: ")
         for timing in self.timings:
             print(f"    - {timing.name}: {format_duration(timing.duration)}")
-        print(f"  - Tasks: {len(self.task_summaries)} - {(len(self.task_summaries) / self.duration.total_seconds()):.1f} tasks/sec")
+        print(
+            f"  - Tasks: {len(self.task_summaries)} - {(len(self.task_summaries) / self.duration.total_seconds()):.1f} tasks/sec"
+        )
+
+        for status in [TaskStatus.FAILED, TaskStatus.FINISHED]:
+            print(
+                f"    - {status.value.lower()}: {len([task for task in self.task_summaries if task.status == status])}"
+            )
         print("--------------------")
 
+    def __enter__(self):
+        return self
 
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        pass
